UI overhaul: card-based mood analysis, audio recording, song previews

+++ backend/emotion/predictor.py
@@ -4,5 +4,8 @@ EMOTIONS = ["Happy", "Sad", "Angry", "Calm", "Neutral"]
 
 def predict_emotion(mfcc, chroma, mel):
     emotion = random.choice(EMOTIONS)
-    confidence = round(random.uniform(0.65, 0.9), 2)
+
+    # Always safe float
+    confidence = float(round(random.uniform(0.65, 0.9), 2))
+
     return emotion, confidence
